+++ lib/mix/tasks/ccxt_extract.update.ex
@@ -207,6 +207,7 @@ defmodule Mix.Tasks.CcxtExtract.Update do
     ccxt_extract.describe
     ccxt_extract.load_markets
     ccxt_extract.url_templates
+    ccxt_extract.request_headers
     ccxt_extract.signing_fixtures
   )